Add "OldInput" functionality for checkbox and radio inputs when array name isset
Example "OldInput" fixed issue:
$builder->checkbox('form_weekday[]')->value('Monday');
$builder->checkbox('form_weekday[]')->value('Tuesday');
$builder->checkbox('form_weekday[]')->value('Wednesday')-;
$builder->checkbox('form_weekday[]')->value('Thursday');
$builder->checkbox('form_weekday[]')->value('Friday');
